api/ml_models.py
import csv
import json

# PyTorch models (will be imported when available)
try:
    import torch
    import torch.nn as nn
    
    class RNNModel(nn.Module):
        def __init__(self, input_size, hidden_size, num_layers):
            super(RNNModel, self).__init__()
            self.rnn = nn.RNN(input_size, hidden_size, num_layers, batch_first=True)
            self.fc = nn.Linear(hidden_size, 1)

        def forward(self, x):
            out, _ = self.rnn(x)
            return self.fc(out[:, -1, :])

    class LSTMModel(nn.Module):
        def __init__(self, input_size, hidden_size, num_layers):
            super(LSTMModel, self).__init__()
            self.lstm = nn.LSTM(input_size, hidden_size, num_layers, batch_first=True)
            self.fc = nn.Linear(hidden_size, 1)

        def forward(self, x):
            out, _ = self.lstm(x)
            return self.fc(out[:, -1, :])

    class GRUModel(nn.Module):
        def __init__(self, input_size, hidden_size, num_layers):
            super(GRUModel, self).__init__()
            self.gru = nn.GRU(input_size, hidden_size, num_layers, batch_first=True)
            self.fc = nn.Linear(hidden_size, 1)

        def forward(self, x):
            out, _ = self.gru(x)
            return self.fc(out[:, -1, :])

    class BiLSTMModel(nn.Module):
        def __init__(self, input_size, hidden_size, num_layers):
            super(BiLSTMModel, self).__init__()
            self.lstm = nn.LSTM(input_size, hidden_size, num_layers, batch_first=True, bidirectional=True)
            self.fc = nn.Linear(hidden_size * 2, 1)

        def forward(self, x):
            out, _ = self.lstm(x)
            return self.fc(out[:, -1, :])

    TORCH_AVAILABLE = True
except ImportError:
    TORCH_AVAILABLE = False
    print("PyTorch not available. Neural network models will not work.")

class FishCastML:
    def __init__(self):
        self.models = {
            'Linear': 'LinearRegression',
        }
        
        if TORCH_AVAILABLE:
            self.models.update({
                'GRU': 'GRUModel',
                'LSTM': 'LSTMModel',
                'BiLSTM': 'BiLSTMModel',
                'RNN': 'RNNModel',
            })
    
    def load_data(self, filepath):
        """Load and preprocess CSV data"""
        # Simple CSV reader for now
        data = []
        with open(filepath, 'r') as file:
            reader = csv.DictReader(file)
            for row in reader:
                data.append(row)
        
        # Convert to simple format for now
        # This is a simplified version - in real implementation you'd want pandas
        return data
    
    def prepare_data(self, df):
        """Prepare data for training"""
        # Simplified version for now
        return None, None, None, None, None
    
    def train_torch_model(self, model_class, X_train, y_train, X_test, y_test, input_size, hidden_size=32, num_layers=1, num_epochs=200, lr=0.001):
        """Train PyTorch model"""
        # Simplified version for now
        return None, 0.0, 0.0
    
    def train_linear_model(self, X_train, y_train, X_test, y_test):
        """Train Linear Regression model"""
        # Simplified version for now
        return None, 0.0, 0.0
    
    def train_and_predict(self, dataset_id, dataset_file, models_to_train=None):
        """Train models and make predictions"""
        if models_to_train is None:
            models_to_train = ['Linear']
        
        # Load data
        data = self.load_data(dataset_file)
        
        # Simplified mock results
        results = {}
        for model_name in models_to_train:
            results[model_name] = {
                'predictions': [1.0, 2.0, 3.0, 4.0, 5.0],
                'actual_values': [1.1, 2.1, 3.1, 4.1, 5.1],
                'mse': 0.01,
                'mae': 0.1
            }
        
        return results
    
    def get_correlation_analysis(self, dataset_file):
        """Generate correlation analysis"""
        data = self.load_data(dataset_file)
        
        # Simplified mock correlation
        corr_matrix = {
            'stok_ikan': {'stok_ikan': 1.0, 'bulan_normalized': 0.5},
            'bulan_normalized': {'stok_ikan': 0.5, 'bulan_normalized': 1.0}
        }
        
        return {
            'correlation_matrix': corr_matrix,
            'plot_base64': None  # No plot for now
        }
    
    def run_optimization(self, dataset_file, population_size=40, generations=100):
        """Run NSGA-III optimization"""
        # Simplified mock optimization
        data = self.load_data(dataset_file)
        
        return {
            'solutions': [[0.1, 0.2, 0.3], [0.4, 0.5, 0.6]],
            'objectives': [[10.0, 0.01], [15.0, 0.02]],
            'best_solution': [0.1, 0.2, 0.3],
            'best_total_stok': 10.0,
            'best_mse': 0.01
        }

# Global instance
ml_engine = FishCastML() 