requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
pandas==2.1.4
numpy==1.26.2

# ML/AI dependencies (optional - for full functionality)
# torch>=2.0.0
# scikit-learn>=1.0.0
# pandas>=1.5.0
# numpy>=1.20.0
# matplotlib>=3.5.0
# seaborn>=0.11.0
# pymoo>=0.6.0 